+++ fpdf/fpdf.py
@@ -401,21 +401,16 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             text (str): HTML content to render
             image_map (function): an optional one-argument function that map <img> "src"
                 to new image URLs
-            li_tag_indent (int): numeric indentation of <li> elements
-            dd_tag_indent (int): numeric indentation of <dd> elements
+            li_tag_indent (int): [**DEPRECATED since v2.7.8**] numeric indentation of <li> elements - Set tag_indents instead
+            dd_tag_indent (int): [**DEPRECATED since v2.7.8**] numeric indentation of <dd> elements - Set tag_indents instead
             table_line_separators (bool): enable horizontal line separators in <table>
             ul_bullet_char (str): bullet character for <ul> elements
             ul_bullet_color (tuple | str | drawing.Device* instance): color of the <ul> bullets
-            heading_sizes (dict): font size per heading level names ("h1", "h2"...)
-            pre_code_font (str): font to use for <pre> & <code> blocks
+            heading_sizes (dict): [**DEPRECATED since v2.7.8**] font size per heading level names ("h1", "h2"...) - Set tag_styles instead
+            pre_code_font (str): [**DEPRECATED since v2.7.8**] font to use for <pre> & <code> blocks - Set tag_styles instead
             warn_on_tags_not_matching (bool): control warnings production for unmatched HTML tags
-            element_colors (dict): dictionary of colors for elements. Possible keys:
-            * "link" sets the link color for a href tag
-            * "li" sets the bullet or number for <li> tags
-            * "blockquote" sets the color for <blockquote> tags
-            * "headings" sets the color is not specified in a h1, h2 etc. heading tag
-
-            The values for the dictionary are a tuple of three ints, being the r,g and b values
+            tag_indents (dict): mapping of HTML tag names to numeric values representing their horizontal left identation
+            tag_styles (dict): mapping of HTML tag names to colors
         """
         kwargs2 = vars(self)
         # Method arguments must override class & instance attributes:
@@ -901,7 +896,7 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             if isinstance(self.page_background, tuple):
                 self.set_fill_color(*self.page_background)
                 self.rect(0, 0, self.w, self.h, style="F")
-                self.set_fill_color(*(255 * v for v in fc.colors))
+                self.set_fill_color(*fc.colors255)
             else:
                 self.image(self.page_background, 0, 0, self.w, self.h)
 
@@ -1111,7 +1106,7 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
             background, (str, io.BytesIO, Image, DeviceRGB, tuple, type(None))
         ):
             if isinstance(background, DeviceRGB):
-                self.page_background = tuple(255 * v for v in background.colors)
+                self.page_background = background.colors255
             else:
                 self.page_background = background
         else:
@@ -4943,7 +4938,11 @@ class FPDF(GraphicsStateMixin, TextRegionMixin):
         prev_font = (self.font_family, self.font_style, self.font_size_pt)
         self.set_font(
             font_face.family or self.font_family,
-            font_face.emphasis.style if font_face.emphasis is not None else "",
+            (
+                font_face.emphasis.style
+                if font_face.emphasis is not None
+                else self.font_style
+            ),
             font_face.size_pt or self.font_size_pt,
         )
         prev_text_color = self.text_color